refractoring code for simplicity. Initial app.py file to handle routing, separated files to handle separate logic for better readability.
Created a webui/ module set:
config.py: env loading, defaults, project paths, favicon helpers.
schemas.py: all request/response models.
model_catalog.py: model directory inspection/discovery/compatibility checks.
model_runtime.py: runtime state, .env persistence, model load/reload lifecycle.
downloads.py: HF download job tracking + worker thread + active job/status APIs.
ui.py: template loading and HTML rendering from index.html.

Moved the inline UI_HTML out of app.py into index.html and keep the same placeholder substitution behavior. Replaced app.py with a thin FastAPI router layer that calls those modules.
