+++ webui/__init__.py
@@ -1,0 +1,2 @@
+"""Web UI backend modules for the AirLLM FastAPI app."""
+
